+++ src/Ext/Techno/Hooks.Misc.cpp
@@ -478,6 +478,25 @@ DEFINE_HOOK(0x6F88BF, TechnoClass_CanAutoTargetObject_AttackMindControlledDelay,
 	return 0;
 }
 
+#pragma region NoTurretUnitAlwaysTurnToTarget
+
+DEFINE_HOOK(0x7410BB, UnitClass_GetFireError_CheckFacingError, 0x8)
+{
+	enum { NoNeedToCheck = 0x74132B, ContinueCheck = 0x7410C3 };
+
+	GET(const FireError, fireError, EAX);
+
+	if (fireError == FireError::OK)
+		return ContinueCheck;
+
+	if (!RulesExt::Global()->UnitWithoutTurretAlwaysTurnToTarget)
+		return NoNeedToCheck;
+
+	GET(UnitClass* const, pThis, ESI);
+
+	return (fireError == FireError::REARM && !pThis->Type->Turret && !pThis->IsWarpingIn()) ? ContinueCheck : NoNeedToCheck;
+}
+
 #pragma endregion
 
 #pragma region Misc