+++ docs/Whats-New.md
@@ -505,6 +505,8 @@ New:
 - Aggressive attack move mission (by CrimRecya)
 - No rearm and reload in EMP or temporal (by CrimRecya)
 - Manually empty and reload ammo (by CrimRecya)
+- No turret unit turn to the target (by CrimRecya & TaranDahl)
+- Sell or undeploy building on impact (by CrimRecya)
 - Unit `Speed` setting now accepts floating point values (by Starkku)
 - Aggressive stance toggle command (auto target unarmed enemy buildings) (by Aephiex)
 - Exclusive SuperWeapon Sidebar (by NetsuNegi & CrimRecya)